+++ HARDWARE/DS18B20/ds18b20.h
@@ -14,7 +14,7 @@
 #define	DS18B20_DQ_IN  PBin(15)  //数据端口	PB15
    	
 u8 DS18B20_Init(void);			//初始化DS18B20
-short DS18B20_Get_Temp(void);	//获取温度
+float DS18B20_Get_Temp(void);	//获取温度
 void DS18B20_Start(void);		//开始温度转换
 void DS18B20_Write_Byte(u8 dat);//写入一个字节
 u8 DS18B20_Read_Byte(void);		//读出一个字节
